Fixed import null bug.
If you tried to import setting file when none existed, it would set every mcm value to none. Even the ones where that wouldn't make sense, so it very badly broke the whole menu.
Now if no file exists it will show a message that there is no file and return, preventing the error.

If a user manually messes with the json and sets values incorrectly, this bug could come back. However, I think that would be not worth the hassle of dealing with unless it suddenly becomes a common issue.