Adds easier settings sharing support for Modlist authors.
This will add the ability to quickly share ostim MCM setting like a regular mo2 or Vortex mod.
This way, rather than having to tell users of a modlist to "Grab these random files and put them in this random directory"
They can just tell the user, "install this mod" and this does the rest.