Fix NPCs trying to perform idles at the start of scenes glitching out
This fix disables ai long enough for them to get into the scene and be stable.

It appears to be working. However, due to the unreliable nature of this bug, more testing is needed.

It also seems to be tight on timing. If the ai is renabled too early, it seems the actor can escape the scene still. Since timing in papyrus being off is usually on the latter-end rather than being too early, I'm hoping this isn't an issue for laggy setups. However, it's hard to say.