OSexIntegrationMain: Match indentation to scope
There were a few areas in OSexIntegrationMain where the scope of the code did not match the indentation, and/or mixed indentation styles were used. This commit unifies the indentation styles so that all indentation uses the tab character (used for the majority of lines) and so that indentation matches scope.