Fixes recursion issue that can occur if ODatabase fails to build.
This is a fix to the Using Native Build Function, Error Database not Found console spam error.
This seems to be primarily caused by ODatabase not existing after it is build with the native function, now it will try using the Native function, then try the papyrus function and then if all else fails it will show a messagebox to report the bug on nexus or discord.

The main cause of this bug seems to be if you try to use OStim without OSex installed, so I added another check on startup for OSex.pex which if it's not detected will show an error like the other startup checks. I also slightly changed the language of the errors to be a little less passive aggressive :laughing:.

In theory, just the change to OSexIntergrationMain.psc is needed as it will cover 98% of the times this issue occurs, with the remainder being caused by permission issues that the ODatabase changes *should* catch. Up to you sai if you want to keep the ODatabase changes or not.